Move orphan-precompiled append from MainRootModule to DummyModule
Mixing DynamicModule into MainRootModule flips the resolve path for every
codegen-generated root to use only moduleDirectChildren (see
core/resolve/src/mill/resolve/ResolveCore.scala:509). That breaks Cross[T]
roots and any other resolve-time machinery that depends on the reflective
branch — observed as regressions in root-cross-module,
precompiled-nested-config-mismatch, module-init-error, os-checker, and the
external-module-aliases example.

The orphan append is only needed when codegen produced no build_.package_
and BuildFileCls falls back to DummyModule. For workspaces with a compiled
root, precompiled siblings are already wired by codegen's per-directory
precompiledChildNames aliases, so no append is needed there.

Move DynamicModule + orphanPrecompiledChildren onto DummyModule itself.
MainRootModule reverts to its pre-branch shape, so codegen-emitted roots
stay on the reflective resolve path and Cross/etc. continue to work.
